+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### 0.6.0
 - Update to match Rump 0.6.0 (pass options)
-- Update `require-all`
+- Update `require-all` and `gulp-autoprefixer`
+- Set up absolute paths for source maps
 
 ### 0.5.0
 - Update to match Rump 0.5.0 (use same minor version)